Improve block db interface
